Things Move
There is now a moving square. I could not figure things out simply, so I
had to make a complicated mess of garbage. Because one class didn't want
to reference an object of the other, only because the object was created
AFTER the class (which shouldn't be a problem in a language like
Python), I had to give the first class a parameter, and through that let
it reference the object. Also, this probably only works on my computer.
I need to get better at PyGame.